Add maven license plugin.
I chose GPL v3 for this project. I like free software and as a software
developer I like the viral part of the GPL too.

Things need some clean up, the artifactId is not very good and the
license inside the Java files is looking pretty bad. But this is the
first step with a license, so I'm going with it.